chore: fixed models
