+++ creditkelin/settings.py
@@ -38,6 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'products',
+    'credit',
 ]
 
 MIDDLEWARE = [
